Random stuff

+++ src/test/java/dev/nincodedo/ninbot/components/stream/banner/steamgriddb/SteamGridDBBannerBuilderTest.java
@@ -20,13 +20,14 @@ import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.fail;
 import static org.instancio.Select.field;
+import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
@@ -47,7 +48,14 @@ class SteamGridDBBannerBuilderTest {
 
     @BeforeAll
     static void before() throws IOException {
-        cache.mkdir();
+        if (cache.exists() && !cache.isDirectory()) {
+            if (!cache.delete()) {
+                fail("Failed to delete non directory cache file");
+            }
+        }
+        if (!cache.mkdir()) {
+            fail("Failed to make cache directory");
+        }
         FileUtils.cleanDirectory(cache);
     }
 
@@ -62,14 +70,10 @@ class SteamGridDBBannerBuilderTest {
         var icon = Path.of("docs", "images", "ninbot-github-logo-small.png").toUri().toURL();
         BaseResponse<Game> gameBaseResponse = Instancio.of(new TypeToken<BaseResponse<Game>>() {
         }).set(field("success"), true).create();
-        GameImage background = Instancio.of(new TypeToken<GameImage>() {
-        }).set(field("url"), "file://" + bg.getPath()).create();
-        GameImage logo = Instancio.of(new TypeToken<GameImage>() {
-        }).set(field("url"), "file://" + icon.getPath()).create();
-        List<GameImage> backgrounds = new ArrayList<>();
-        backgrounds.add(background);
-        List<GameImage> logos = new ArrayList<>();
-        logos.add(logo);
+        List<GameImage> backgrounds = Instancio.of(new TypeToken<GameImage>() {
+        }).set(field("url"), "file://" + bg.getPath()).stream().limit(5).toList();
+        List<GameImage> logos = Instancio.of(new TypeToken<GameImage>() {
+        }).set(field("url"), "file://" + icon.getPath()).stream().limit(5).toList();
         BaseResponse<GameImage> logoResponse = Instancio.of(new TypeToken<BaseResponse<GameImage>>() {
         }).set(field("success"), true).set(field("data"), logos).create();
         BaseResponse<GameImage> heroResponse = Instancio.of(new TypeToken<BaseResponse<GameImage>>() {
@@ -87,9 +91,10 @@ class SteamGridDBBannerBuilderTest {
         assertThat(cache).isNotEmptyDirectory();
         String bannerFileName = cache.listFiles()[0].getName();
         assertThat(bannerFileName).contains("Zeldo");
+        assertThat(banner.getFile()).exists();
 
         banner.setFile(null);
-        when(gameBannerRepository.findGameBannerByLogoIdAndBackgroundId(banner.getLogoId(), banner.getBackgroundId()))
+        when(gameBannerRepository.findGameBannerByLogoIdAndBackgroundId(anyInt(), anyInt()))
                 .thenReturn(Optional.of(banner));
 
         verify(steamGridDBFeign).searchGameByName(anyString());
@@ -99,7 +104,11 @@ class SteamGridDBBannerBuilderTest {
         var cachedBannerAsync = steamGridDBBannerBuilder.getGameBannerAsync("Zeldo");
         var cachedBanner = cachedBannerAsync.get();
 
-        assertThat(cachedBanner.getFile()).hasName(bannerFileName);
+        assertThat(cachedBanner.getFile().getName()).contains("Zeldo");
+
+        var bannerFromFileOptional = steamGridDBBannerBuilder.getGameBannerFromFile(cachedBanner.getFile());
+
+        assertThat(bannerFromFileOptional).isPresent().contains(banner);
 
         verifyNoMoreInteractions(steamGridDBFeign);
     }
